+++ lib/project_types/theme/commands/pull.rb
@@ -34,10 +34,13 @@ module Theme
           flags[:ignores] ||= []
           flags[:ignores] |= pattern
         end
+        parser.on("-f", "--force") { flags[:force] = true }
       end
 
       def call(_args, name)
         root = root_value(options, name)
+        return if exist_and_not_empty?(root) && !valid_theme_directory?(root)
+
         delete = !options.flags[:nodelete]
         theme = find_theme(root, **options.flags)
         return if theme.nil?